Move BasicData#many integration test to integration folder
